Add typing to dataset objects, move recursive_map
