+++ espei/core_utils.py
@@ -2,18 +2,19 @@
 Utilities for querying and modifiying datasets.
 """
 import copy
+from typing import List
 import numpy as np
 import tinydb
+from espei.datasets import Dataset
 from espei.sublattice_tools import canonicalize, recursive_tuplify
 
-
-def filter_configurations(desired_data, configuration, symmetry):
+def filter_configurations(desired_data: List[Dataset], configuration, symmetry) -> List[Dataset]:
     """
     Return non-equilibrium thermochemical datasets with invalid configurations removed.
 
     Parameters
     ----------
-    desired_data : List[Dict[str, Any]]
+    desired_data : List[Dataset]
         List of non-equilibrium thermochemical datasets
     configuration : tuple
         Sublattice configuration as a tuple, e.g. ("CU", ("CU", "MG"))
@@ -22,7 +23,7 @@ def filter_configurations(desired_data, configuration, symmetry):
 
     Returns
     -------
-    List[Dict[str, Any]]
+    List[Dataset]
 
     """
     for data in desired_data:
@@ -38,7 +39,7 @@ def filter_configurations(desired_data, configuration, symmetry):
     return desired_data
 
 
-def filter_temperatures(desired_data):
+def filter_temperatures(desired_data) -> List[Dataset]:
     """
     Return non-equilibrium thermochemical datasets with temperatures below 298.15 K removed.
 
@@ -50,12 +51,12 @@ def filter_temperatures(desired_data):
 
     Parameters
     ----------
-    desired_data : List[Dict[str, Any]]
+    desired_data : List[Dataset]
         List of non-equilibrium thermochemical datasets
 
     Returns
     -------
-    List[Dict[str, Any]]
+    List[Dataset]
 
     """
     for data in desired_data:
@@ -223,33 +224,6 @@ def ravel_zpf_values(desired_data, independent_comps, conditions=None):
     return equilibria_dict
 
 
-def recursive_map(f, x):
-    """
-    map, but over nested lists
-
-    Parameters
-    ----------
-    f : callable
-        Function to apply to x
-    x : list or value
-        Value passed to v
-
-    Returns
-    -------
-    list or value
-    """
-    if isinstance(x, list):
-        if [isinstance(xx, list) for xx in x]:
-            # we got a nested list
-            return [recursive_map(f, xx) for xx in x]
-        else:
-            # it's a list with some values inside
-            return list(map(f, x))
-    else:
-        # not a list, probably just a singular value
-        return f(x)
-
-
 def symmetry_filter(x, config, symmetry):
     """
     Return True if the candidate sublattice configuration has any symmetry
@@ -284,7 +258,7 @@ def symmetry_filter(x, config, symmetry):
     return False
 
 
-def get_prop_data(comps, phase_name, prop, datasets, additional_query=None):
+def get_prop_data(comps, phase_name, prop, datasets, additional_query=None) -> List[Dataset]:
     """
     Return a copy of datasets that match the components, phase and property.
 
@@ -305,8 +279,7 @@ def get_prop_data(comps, phase_name, prop, datasets, additional_query=None):
 
     Returns
     -------
-    list
-        List of dictionary datasets that match the criteria
+    List[Dataset]
 
     """
     if additional_query is None:
